refactor: changed uart_demo by lora_app
